refactor: grpc logic to handler (#5731)
