+++ jina/serve/runtimes/gateway/grpc/gateway.py
@@ -1,18 +1,12 @@
-from typing import TYPE_CHECKING, AsyncIterator, Optional
+from typing import Optional
 
 import grpc
 from grpc_health.v1 import health, health_pb2, health_pb2_grpc
 from grpc_reflection.v1alpha import reflection
 
-from jina.helper import get_full_version
 from jina.proto import jina_pb2, jina_pb2_grpc
 from jina.serve.gateway import BaseGateway
 from jina.serve.runtimes.helper import _get_grpc_server_options
-from jina.types.request.data import DataRequest
-from jina.types.request.status import StatusMessage
-
-if TYPE_CHECKING:  # pragma: no cover
-    from jina.types.request import Request
 
 
 class GRPCGateway(BaseGateway):
@@ -46,12 +40,12 @@ class GRPCGateway(BaseGateway):
             interceptors=self.grpc_tracing_server_interceptors,
         )
 
-        jina_pb2_grpc.add_JinaRPCServicer_to_server(self, self.server)
+        jina_pb2_grpc.add_JinaRPCServicer_to_server(self._request_handler, self.server)
 
-        jina_pb2_grpc.add_JinaSingleDataRequestRPCServicer_to_server(self, self.server)
+        jina_pb2_grpc.add_JinaSingleDataRequestRPCServicer_to_server(self._request_handler, self.server)
 
-        jina_pb2_grpc.add_JinaGatewayDryRunRPCServicer_to_server(self, self.server)
-        jina_pb2_grpc.add_JinaInfoRPCServicer_to_server(self, self.server)
+        jina_pb2_grpc.add_JinaGatewayDryRunRPCServicer_to_server(self._request_handler, self.server)
+        jina_pb2_grpc.add_JinaInfoRPCServicer_to_server(self._request_handler, self.server)
 
         service_names = (
             jina_pb2.DESCRIPTOR.services_by_name['JinaRPC'].full_name,
@@ -105,73 +99,3 @@ class GRPCGateway(BaseGateway):
     async def run_server(self):
         """Run GRPC server forever"""
         await self.server.wait_for_termination()
-
-    async def dry_run(self, empty, context) -> jina_pb2.StatusProto:
-        """
-        Process the call requested by having a dry run call to every Executor in the graph
-
-        :param empty: The service expects an empty protobuf message
-        :param context: grpc context
-        :returns: the response request
-        """
-        from jina._docarray import Document, DocumentArray
-        from jina.serve.executors import __dry_run_endpoint__
-
-        da = DocumentArray([Document()])
-        try:
-            async for _ in self.streamer.stream_docs(
-                docs=da, exec_endpoint=__dry_run_endpoint__, request_size=1
-            ):
-                pass
-            status_message = StatusMessage()
-            status_message.set_code(jina_pb2.StatusProto.SUCCESS)
-            return status_message.proto
-        except Exception as ex:
-            status_message = StatusMessage()
-            status_message.set_exception(ex)
-            return status_message.proto
-
-    async def _status(self, empty, context) -> jina_pb2.JinaInfoProto:
-        """
-        Process the the call requested and return the JinaInfo of the Runtime
-
-        :param empty: The service expects an empty protobuf message
-        :param context: grpc context
-        :returns: the response request
-        """
-        info_proto = jina_pb2.JinaInfoProto()
-        version, env_info = get_full_version()
-        for k, v in version.items():
-            info_proto.jina[k] = str(v)
-        for k, v in env_info.items():
-            info_proto.envs[k] = str(v)
-        return info_proto
-
-    async def stream(
-        self, request_iterator, context=None, *args, **kwargs
-    ) -> AsyncIterator['Request']:
-        """
-        stream requests from client iterator and stream responses back.
-
-        :param request_iterator: iterator of requests
-        :param context: context of the grpc call
-        :param args: positional arguments
-        :param kwargs: keyword arguments
-        :yield: responses to the request after streaming to Executors in Flow
-        """
-        async for resp in self.streamer.rpc_stream(
-            request_iterator=request_iterator, context=context, *args, **kwargs
-        ):
-            yield resp
-
-    async def process_single_data(
-        self, request: DataRequest, context=None
-    ) -> DataRequest:
-        """Implements request and response handling of a single DataRequest
-        :param request: DataRequest from Client
-        :param context: grpc context
-        :return: response DataRequest
-        """
-        return await self.streamer.process_single_data(request, context)
-
-    Call = stream
